Simplify PHPUnit configuration, add new key to blockade config, rename model and trait, refactor ExludeBlockedModel scope
A considerable refactoring has been performed, reducing complexity in several files. The phpunit.xml.dist was simplified for better readability. A user_foreign_key configuration option was added to config/blockade.php to explicit declare user model's foreign key. TestModel has been renamed to User to better reflect its purpose. The ExcludeBlockedModels trait was renamed to HasBlocked for clearer intent. This trait was removed from CanBlock to avoid duplication. The ExcludeBlockedModelsScope was refactored to utilize the new user_foreign_key option. RefreshDatabase trait usage was moved to a more global level for DRY purposes. This was all done to decrease the level of complexity and increase readability and maintainability of the code.