Simplify test workflow and introduce linting
The current test workflow was split into separate workflows for tests and linting.
The 'run-tests.yml' workflow was deleted and, in its place; 'run-pest.yml' and 'run-linter.yml' files were created.
This restructuring facilitates easier maintenance and better visibility in case of failure.
In addition to that, a new linting job was introduced to check code styling inconsistencies across the project.